+++ zvtm/trunk/src/main/java/fr/inria/zvtm/engine/ObservedRegionListener.java
@@ -2,7 +2,7 @@
  *   DATE OF CREATION:  Wed Jul 12 15:01:11 2006
  *   AUTHOR :           Emmanuel Pietriga (emmanuel.pietriga@inria.fr)
  *   MODIF:             Emmanuel Pietriga (emmanuel.pietriga@inria.fr)
- *   Copyright (c) INRIA, 2006. All Rights Reserved
+ *   Copyright (c) INRIA, 2006-2010. All Rights Reserved
  *   Licensed under the GNU LGPL. For full terms see the file COPYING.
  *
  * $Id$
@@ -17,7 +17,7 @@ package fr.inria.zvtm.engine;
 public interface ObservedRegionListener {
 
     /**projected values of west, north, east and south intersections*/
-    public void intersectsParentRegion(long[] wnes);
+    public void intersectsParentRegion(double[] wnes);
     
 
 }
